[FEAT] uwb_init::feed_uwb(): outlier reject + closest point + subsample
HINT: feed_uwb() and feed_position() allows to restrict the number of
samples. If threshold is exceeded the buffer is subsampled by 2.
Position updates are only considered if a position displacement according
to a threshold (min_traveld_distance) happened.
feed_uwb() checks if a new closest point to the anchor was obtained.